% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{as.factors}
\alias{as.factors}
\title{Factorize variables}
\usage{
as.factors(data, tofactor)
}
\arguments{
\item{data}{The data frame}

\item{tofactor}{A vector indicating the column names of the column which should be factorized.}
}
\value{
Data frame
}
\description{
Indicate columns of a data frame as factors
}
