% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{ezAnalysis}
\alias{ezAnalysis}
\title{Compute ANOVA, LMM, or t-test}
\usage{
ezAnalysis(data, dv, wid = NULL, within = NULL, between = NULL,
  test = "aov", options = list(random_effects = NULL, contrasts = NULL, ddf
  = "Satterthwaite", sph.cor = "GG", mau.p = 0.05, peta = TRUE),
  return_obj = FALSE, print = TRUE, dfsep = ", ")
}
\arguments{
\item{data}{Data frame containing the data to be analyzed.}

\item{dv}{Name of the column in data that contains the dependent variable. Values in this column must be numeric.}

\item{wid}{Name of the column in data that contains the variable specifying the case/Ss identifier. This should be a unique value per case/Ss.}

\item{within}{Names of columns in data that contain predictor variables that are manipulated (or observed) within-Ss. If a single value, may be specified by name alone; if multiple values, must be specified as a .() list.}

\item{between}{Names of columns in data that contain predictor variables that are manipulated (or observed) between-Ss. If a single value, may be specified by name alone; if multiple values, must be specified as a .() list.}

\item{test}{Use this parameter to specify the test that should be run. (aov = ANOVA, lmer = linear mixed effects model)}

\item{options}{Use this parameter to specify the test options. For lmer these are random_effects, contrasts and ddf. For aov sph.cor and mau.p}

\item{return_obj}{Set this parameter to TRUE if you want that the analysis model is returned as function value.}

\item{print}{Set this parameter to FALSE if the function should not print the results.}

\item{dfsep}{This specifies the seperator used to seperator the df values in the formatted results. By default, a comma is used.}

\item{random_effects}{LMER parameter: Use this parameter to specify a more complex random effects term of a linear mixed effects model. The given value must be of type string.}

\item{contrasts}{LMER parameter: Use this paramater to specify a list of contrasts for the fixed effects factors in a linear mixed effects model.}

\item{ddf}{LMER parameter: This specifies the method used for calculating p-values of the fixed effects of the linear mixed model (Satterthwaite, Kenward-Roger, etc.).}

\item{sph.cor}{ANOVA parameter: Use this paramater to specify the correction estimates to use for sphericity corrections of within factors of a ANOVA (GG, HF, no; default="GG").}

\item{mau.p}{ANOVA parameter: Use this paramater to specify the threshold for Mauchly's test of sphericity (default=0.05).}
}
\value{
Model object (optional)
}
\description{
This function provides easy analysis of data from factorial experiments,
including purely within-Ss designs (a.k.a. "repeated measures"),
purely between-Ss designs, and mixed within-and-between-Ss designs,
yielding either Linear mixed effect model, ANOVA or t-test results,
as well as effect sizes and assumption checks.
}
