% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{lsmeans_out}
\alias{lsmeans_out}
\title{Distilles the most relevant data from an output object of lsmeans, calculates effect sizes, and displays the results in a compact format.}
\usage{
lsmeans_out(lsmobj, paired = FALSE, n.equal = TRUE)
}
\arguments{
\item{lsmobj}{Output object created by a call to lsmeans.}

\item{paired}{Set to true for paired data}

\item{n.equal}{Set to false if number of subjects per group is not equal}
}
\description{
Distilles the most relevant data from an output object of lsmeans, calculates effect sizes, and displays the results in a compact format.
}
