% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot}
\usage{
multiplot(..., plotlist = NULL, shared_legend = FALSE, cols = 1)
}
\arguments{
\item{...}{Plots}

\item{plotlist}{List of plots}

\item{shared_legend}{Set to true if plots should share the same legend}

\item{cols}{Number of cols of the shared legend plot}
}
\description{
Plot muliple plots into a single plot. Optionally, it is possible that plots share a single legend (of first plot in provided)
}
