% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{normDataWithin}
\alias{normDataWithin}
\title{Normalize within-group data}
\usage{
normDataWithin(data = NULL, idvar, measurevar, betweenvars = NULL,
  na.rm = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{idvar}{The name of a column that identifies each subject (or matched subjects)#'}

\item{measurevar}{The name of a column that contains the variable to be summariezed}

\item{betweenvars}{A vector containing names of columns that are between-subjects variables}

\item{na.rm}{Boolean that indicates whether to ignore NA's}

\item{conf.interval}{The percent range of the confidence interval (default is 95%)}
}
\value{
a data frame with normalized data
}
\description{
Norms the data within specified groups in a data frame; it normalizes each subject (identified by idvar) so that they have the same mean, within each group specified by betweenvars.
}
