% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.aov.residuals}
\alias{plot.aov.residuals}
\title{Plot AOV Residuals}
\usage{
\method{plot}{aov.residuals}(aov)
}
\arguments{
\item{aov}{AOV object}
}
\description{
Plots the residuals of a ANOVA (aov) object for inspection
}
