% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotresiduals.aov}
\alias{plotresiduals.aov}
\title{Plot AOV Residuals}
\usage{
plotresiduals.aov(aov)
}
\arguments{
\item{aov}{AOV object}
}
\description{
Plots the residuals of an ANOVA (aov) object for inspection
}
