% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework.R
\name{plots.add}
\alias{plots.add}
\title{Add to Plots}
\usage{
plots.add(plot, name, id = 0)
}
\arguments{
\item{plot}{The plot}

\item{name}{Name of the plot}

\item{id}{Plot id (optional)}
}
\value{
The plot
}
\description{
Add a plot to the experiments' plots tibble
}
