% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework.R
\name{results.add}
\alias{results.add}
\title{Add to Results}
\usage{
results.add(result, name, type = "", id = 0)
}
\arguments{
\item{result}{The result}

\item{name}{Name of the result}

\item{type}{The type of the results (e.g., Correlation, t-test, ANOVA)}

\item{id}{Result id (optional)}
}
\value{
The result
}
\description{
Add any result to the results tibble of the experiment
}
