% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework.R
\name{results.get}
\alias{results.get}
\title{Get Result}
\usage{
results.get(id)
}
\arguments{
\item{id}{The result id or name}
}
\value{
The result
}
\description{
Get a result from the experiments' results tibble
}
