% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{rm.outlier}
\alias{rm.outlier}
\title{Remove Outliers in Data Frame}
\usage{
rm.outlier(data, columns, within_subjects = FALSE, wid = NULL,
  median = FALSE, sd_times = 3, print = FALSE)
}
\arguments{
\item{data}{The data frame}

\item{columns}{Dependent variables}

\item{within_subjects}{Remove outlier within subjects or overall subjects?}

\item{wid}{Subject identifier variable}

\item{median}{Should the outliers be identified via standard deviation or via median (boxplot)?}

\item{sd_times}{Below/above how many standard deviations from the mean is a value considered as an outlier?}

\item{print}{Should the proportion of outliers in the data be printed?}
}
\value{
Data frame with removed outliers (indicated as NA)
}
\description{
Removes outliers in a given data frame. Outliers can be remove within or across subjects.
}
