% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{se}
\alias{se}
\title{Calculates the standard error of the mean for a given vector}
\usage{
se(x)
}
\arguments{
\item{x}{Data (vector)}
}
\value{
The standard error
}
\description{
Calculates the standard error of the mean for a given vector
}
