% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework.R
\name{subsets.add}
\alias{subsets.add}
\title{Add to Data Subsets}
\usage{
subsets.add(subset, name, id = 0)
}
\arguments{
\item{subset}{Data}

\item{name}{Name of the subset}

\item{id}{Subset id (optional)}
}
\value{
The subset
}
\description{
Add a tibble or data frame to the experiment data subsets
}
