% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework.R
\name{subsets.get}
\alias{subsets.get}
\title{Get Subset}
\usage{
subsets.get(id)
}
\arguments{
\item{id}{The subset id or name}
}
\value{
The subset
}
\description{
Get a subset from the experiment data subsets
}
