% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{summarySE}
\alias{summarySE}
\title{Gives count, mean, standard deviation, standard error of the mean, and confidence interval (default 95%).}
\usage{
summarySE(data = NULL, measurevar, groupvars = NULL, na.rm = TRUE,
  conf.interval = 0.95)
}
\arguments{
\item{data}{A data frame}

\item{measurevar}{The name of a column that contains the variable to be summariezed}

\item{groupvars}{Vector containing names of columns that contain grouping variables}

\item{na.rm}{Boolean that indicates whether to ignore NA's}

\item{conf.interval}{The percent range of the confidence interval (default is 95%)}
}
\value{
data.frame   A data frame with count, mean, standard deviation, standard error of the mean, and confidence interval (default 95%).
}
\description{
Gives count, mean, standard deviation, standard error of the mean, and confidence interval (default 95%).
}
