% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{summarySEwithin}
\alias{summarySEwithin}
\title{Summarize within-subjects data}
\usage{
summarySEwithin(data = NULL, measurevar, betweenvars = NULL,
  withinvars = NULL, idvar = NULL, na.rm = TRUE, conf.interval = 0.95)
}
\arguments{
\item{data}{A data frame}

\item{measurevar}{The name of a column that contains the variable to be summariezed}

\item{betweenvars}{A vector containing names of columns that are between-subjects variables}

\item{withinvars}{Vector containing names of columns that are within-subjects variables}

\item{idvar}{The name of a column that identifies each subject (or matched subjects)}

\item{na.rm}{Boolean that indicates whether to ignore NA's}

\item{conf.interval}{The percent range of the confidence interval (default is 95%)}
}
\value{
a data frame with count, mean, standard deviation, standard error of the mean, and confidence interval (default 95%).
}
\description{
Summarizes data, handling within-subjects variables by removing inter-subject variability
}
