% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_apa}
\alias{theme_apa}
\title{APA Theme for ggplot}
\usage{
theme_apa(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\value{
The theme
}
\description{
Theme for ggplot following APA guidelines
}
\examples{
p <- ggplot(mtcars) + geom_point(aes(x = wt, y = mpg, colour = factor(gear))) + facet_wrap(~am)
p + theme_apa()
}
