% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_apa}
\alias{theme_apa}
\title{APA Theme for ggplot}
\usage{
theme_apa(plot.box = FALSE, base_size = 20)
}
\arguments{
\item{plot.box}{Is your plot a boxplot? (boolean)}

\item{font_size_multiplier}{By how much do you want to muliply for font size? (Standard is 1)}
}
\value{
The theme
}
\description{
Theme for ggplot following APA guidelines
}
